fixed a warning

+++ base/ipc/sem/sem.c
@@ -1446,9 +1446,9 @@ RTAI_SYSCALL_MODE int rt_spl_lock_timed(SPL *spl, unsigned long ns)
 		spl->count++;
 	} else {
 		RTIME end_time;
-		int locked;
+		long locked;
 		end_time = rdtsc() + imuldiv(ns, tuned.cpu_freq, 1000000000);
-		while ((locked = cmpxchg(&spl->owndby, 0L, rt_current)) && rdtsc() < end_time);
+		while ((locked = (long)cmpxchg(&spl->owndby, 0L, rt_current)) && rdtsc() < end_time);
 		if (locked) {
 			rtai_restore_flags(flags);
 			return -1;
